Fix some problems with accept queue maintenance
Accepted sockets that were reset before making it to the established state were
not removing their hold on an entry in the accept queue. This change modifies
the state change callbacks to include the old state, allowing the closing
callback to be used by the parent socket to free the entry in the queue.