Allow the conversion from Sockets to Handles, and allow users to test whether Sockets are ready for reads or writes.
This commit allows user to perform non-blocking testing of whether or not a
Socket can read or write. This is useful for a number of reasons, but is also
a necessary component for allowing for the conversion to GHC Handles. This
conversion also requires Sockets to now be members of Typeable.

(This is a squash merge of the 'can-readwrite' branch.)